Move protobuf reference class to its own header
Should be useful for other tests, put it in a common location.
